second Commit

+++ app/src/main/java/com/example/garageapp/data/model/CounterEntity.kt
@@ -3,6 +3,6 @@ package com.example.garageapp.data.model
 data class CounterEntity(
     val counterId: String = "",
     val shopId: String = "",
-    val jobCardNextNumber: Int = 1,
-    val invoiceNextNumber: Int = 1
+    val jobCardNextNumber: Long = 1L,
+    val invoiceNextNumber: Long = 1L
 )
