+++ app/src/main/java/com/example/garageapp/domain/model/Counter.kt
@@ -3,6 +3,6 @@ package com.example.garageapp.domain.model
 data class Counter(
     val counterId: String = "",
     val shopId: String = "",
-    val jobCardNextNumber: Int = 1,
-    val invoiceNextNumber: Int = 1
+    val jobCardNextNumber: Long = 1L,
+    val invoiceNextNumber: Long = 1L
 )
